Do not delete src_image if --notest_mod is set.
TEST=From an already built image dir, ran archive_build with --notest_mod
set and verified the image had chromiumos_base_image.bin and chromiumos_image.bin.

Review URL: http://codereview.chromium.org/2111015

+++ archive_build.sh
@@ -149,8 +149,10 @@ then
       "--factory_install"
 fi
 
-# Remove the developer build.
-rm -f ${SRC_IMAGE}
+# Remove the developer build if test image is also built.
+if [ $FLAGS_test_mod -eq $FLAGS_TRUE ] ; then
+  rm -f ${SRC_IMAGE}
+fi
 
 # Zip the build
 echo "Compressing and archiving build..."
